+++ src/shapes/IText/ITextBehavior.ts
@@ -406,8 +406,10 @@ export abstract class ITextBehavior<
     this.fire('editing:entered', e ? { e } : undefined);
     this._fireSelectionChanged();
     if (this.canvas) {
-      // @ts-expect-error in reality it is an IText instance
-      this.canvas.fire('text:editing:entered', { target: this, e });
+      this.canvas.fire('text:editing:entered', {
+        target: this as unknown as IText,
+        e,
+      });
       this.canvas.requestRenderAll();
     }
   }