+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [next]
 
+- chore(TS): Update TS to latest [#10044](https://github.com/fabricjs/fabric.js/pull/10044)
 - feat(): Add has method to classRegistry to allow to check if a class exists. (fixes #10001)
 
 ## [6.1.0]